+++ app/Jobs/PlatformStatsSummaryJob.php
@@ -3,14 +3,12 @@
 namespace App\Jobs;
 
 use App\Wiki;
+use App\User;
 use Illuminate\Database\DatabaseManager;
 use PDO;
 use Carbon\Carbon;
 use Illuminate\Support\Facades\Log;
-use Illuminate\Notifications\Notifiable;
-use App\Notifications\PlatformStatsSummaryNotification;
 use Illuminate\Support\Facades\Config;
-use Illuminate\Support\Facades\Notification;
 use Illuminate\Support\Facades\App;
 
 /*
@@ -30,24 +28,39 @@ use Illuminate\Support\Facades\App;
 class PlatformStatsSummaryJob extends Job
 {
     private $inactiveThreshold;
+    private $creationRateRanges;
 
     private $platformSummaryStatsVersion = "v1";
     public function __construct() {
         $this->inactiveThreshold = Config::get('wbstack.platform_summary_inactive_threshold');
+        $this->creationRateRanges = Config::get('wbstack.platform_summary_creation_rate_ranges');
     }
 
     private function isNullOrEmpty( $value ): bool {
         return is_null($value) || intVal($value) === 0;
     }
 
-    public function prepareStats( array $allStats, $wikis ): array {
+    public function getCreationStats(): array {
+        $result = [];
+        $now = Carbon::now();
+        foreach ($this->creationRateRanges as $range) {
+            $limit = $now->clone()->sub(new \DateInterval($range));
+            $wikis = Wiki::where('created_at', '>=', $limit)->count();
+            $result['wikis_created_'.$range] = $wikis;
+            $users = User::where('created_at', '>=', $limit)->count();
+            $result['users_created_'.$range] = $users;
+        }
+        return $result;
+    }
+
+    public function prepareStats( array $allStats, $wikis): array {
 
         $deletedWikis = [];
-        $activeWikis=  [];
-        $inactive =  [];
-        $emptyWikis =  [];
-
+        $activeWikis = [];
+        $inactive = [];
+        $emptyWikis = [];
         $nonDeletedStats = [];
+
         $currentTime = Carbon::now()->timestamp;
 
         foreach( $wikis as $wiki ) {
@@ -60,7 +73,6 @@ class PlatformStatsSummaryJob extends Job
             $wikiDb = $wiki->wikiDb()->first();
 
             if( !$wikiDb ) {
-
                 Log::error(__METHOD__ . ": Could not find WikiDB for {$wiki->domain}");
                 continue;
             }
@@ -71,7 +83,7 @@ class PlatformStatsSummaryJob extends Job
                 Log::warning(__METHOD__ . ": Could not find stats for {$wiki->domain}");
                 continue;
             }
-            
+
             $stats = $allStats[$found_key];
 
             // is it empty?
@@ -86,7 +98,7 @@ class PlatformStatsSummaryJob extends Job
             if(!is_null($stats['lastEdit'])){
                 $lastTimestamp = intVal($stats['lastEdit']);
                 $diff = $currentTime - $lastTimestamp;
-                
+
                 if ($diff >= $this->inactiveThreshold) {
                     $inactive[] = $wiki;
                     continue;
@@ -95,7 +107,7 @@ class PlatformStatsSummaryJob extends Job
 
             $activeWikis[] = $wiki;
         }
-        
+
         $totalNonDeletedUsers = array_sum(array_column($nonDeletedStats, 'users'));
         $totalNonDeletedActiveUsers = array_sum(array_column($nonDeletedStats, 'active_users'));
         $totalNonDeletedPages = array_sum(array_column($nonDeletedStats, 'pages'));
@@ -133,11 +145,11 @@ class PlatformStatsSummaryJob extends Job
         $mediawikiPdo = $mwConn->getPdo();
 
         $pdo->setAttribute(PDO::ATTR_EMULATE_PREPARES, 1);
-        
+
         // prepare the first query
         $statement = $pdo->prepare($this->wikiStatsQuery);
         $statement->execute();
-        
+
         // produces the stats query
         $result = $statement->fetchAll(PDO::FETCH_ASSOC)[0];
         $query = array_values($result)[0];
@@ -146,9 +158,12 @@ class PlatformStatsSummaryJob extends Job
         $allStats = $mediawikiPdo->query($query)->fetchAll(PDO::FETCH_ASSOC);
         $summary = $this->prepareStats( $allStats, $wikis );
 
+        $creationStats = $this->getCreationStats();
+        $summary = array_merge($summary, $creationStats);
+
         $manager->purge('mw');
         $manager->purge('mysql');
-        
+
         // Output to be scraped from logs
         if( !App::runningUnitTests() ) {
             print( json_encode($summary) . PHP_EOL );
@@ -183,7 +198,7 @@ SELECT '",wikis.domain,"' as wiki
 "
 
 ) SEPARATOR ' UNION ALL ')
-    
+
     FROM apidb.wiki_dbs
     LEFT JOIN apidb.wikis ON wiki_dbs.wiki_id = wikis.id;